agentcore-tutorial/pyproject.toml
[project]
name = "agent-deployment"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "bedrock-agentcore>=1.1.1",
    "strands-agents>=1.20.0",
    "strands-agents-tools>=0.2.18",
]
